feat : 최종

+++ src/main/java/com/sparta/todo/service/LikeService.java
@@ -1,11 +1,7 @@
 package com.sparta.todo.service;
 
-import com.sparta.todo.dto.SuccessMessageDto;
-import com.sparta.todo.dto.request.PostRequestDto;
-import com.sparta.todo.dto.request.ToDoRequestDto;
-import com.sparta.todo.dto.response.LikeResponseDto;
-import com.sparta.todo.dto.response.PostResponseDto;
-import com.sparta.todo.dto.response.ToDoResponseDto;
+import com.sparta.todo.dto.requestDto.PostRequestDto;
+import com.sparta.todo.dto.responseDto.PostResponseDto;
 import com.sparta.todo.entity.LikePost;
 import com.sparta.todo.entity.Post;
 import com.sparta.todo.entity.User;
@@ -15,8 +11,6 @@ import com.sparta.todo.repository.LikePostRepository;
 import com.sparta.todo.repository.PostRepository;
 import com.sparta.todo.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,7 +24,7 @@ public class LikeService {
 
     private final UserRepository userRepository;
 
-    /* Post(하루 일정) 좋아요 기능 */
+    // Post(하루 일정) 좋아요 기능
     @Transactional
     public PostResponseDto likePost(Long postId, PostRequestDto postRequestDto, User user) {
 
@@ -45,7 +39,7 @@ public class LikeService {
         if(found.isEmpty()&&postRequestDto.getLikeStatus().equals(true)){
             likePostRepository.save(new LikePost(user, post));
             post.update(true, likeCnt+1);
-        } else if (!found.isEmpty()) {
+        } else if (!found.isEmpty()&&postRequestDto.getLikeStatus().equals(false)) {
             likePostRepository.delete(found.get());
             likePostRepository.flush();
             post.update(false, likeCnt-1);
